Fix incorrect scroll-height of "Clients" page on IE
Possibly a miscalculation due to the client-grid's "flow vertically then
rotate -90deg upright to give the impression the cells flow horizontally
in columns" effect. Other browsers calculated the transformed dimensions
just fine, but IE seemed to be having troubles reading them; meaning the
rows were still flowing vertically down the page as far as its box-model
is concerned.

See line #128 of page.clients.css (relative to this commit) for a better
explanation of what the hell's going on.