Use CSS3 flexboxes for client-grid
As mentioned in dfe9f99, flexboxes turned out to be the perfect solution
for having the logos flow sideways. Because the flexible box model isn't
supported by IE9 in any form, we're ditching support for it in favour of
graceful degradation instead: the grid will flow vertically, without any
hover-revealed details.